Consolidate parse/ into three logical files
- source.go: Source type + Read (raw input concerns together)
- format.go: Format interface + PluginSpec (unchanged)
- pipeline.go: Binding, Bindings, LiteralBindings, Pipeline, Result,
  ErrNoValue, NewResult (everything describing parse output)

Removes the one-typedef source.go, standalone read.go, and result.go.
